Add optional error output port to SendMail
