+++ com.ibm.streamsx.mail/impl/java/src/com/ibm/streamsx/mail/messages.properties
@@ -18,6 +18,7 @@ LOG_ERROR_CAN_NOT_SEND=CDIST3706E Can not send E-mail. smtp host: {0}:{1} to: {2
 # TRNOTE - Do not translate smtp host
 LOG_ERROR_RECEIVE=CDIST3707E Can not receive E-mail. imap host: {0}:{1} username: {2} folder: {3}
 # TRNOTE - Do not translate imap host
+LOG_WARN_MIME_TYPE=CDIST3730W "Content type: {0} is not supported/ignored in email imapHost: {1}:{2} from: {3} date: {4}
 LOG_INFO_MAIL_SEND=CDIST3740I E-mail was sent to: {0} subject: {1}
 
 #END
